edit MainActivity

+++ RssiService/src/com/kiss/math/QuadraticEquation.java
@@ -2,11 +2,11 @@ package com.kiss.math;
 
 
 class QuadraticEquation {
-    private float a;
-    private float b;
-    private float c;
-    private float discriminant;
-    QuadraticEquation(float newA, float newB, float newC) throws Exception {
+    private double a;
+    private double b;
+    private double c;
+    private double discriminant;
+    QuadraticEquation(double newA, double newB, double newC) throws Exception {
         if(newA ==0 ){
             throw new Exception("a=0; that is not a quadratic equation");
         }
@@ -16,13 +16,13 @@ class QuadraticEquation {
         discriminant = (b*b - 4 * a * c);
     }
     
-    public float getDiscriminant() {
+    public double getDiscriminant() {
         return discriminant;
     }
-    public float getV1(){
-        return (float) (-b + Math.sqrt(getDiscriminant()))/(2*a);
+    public double getV1(){
+        return (double) (-b + Math.sqrt(getDiscriminant()))/(2*a);
     }
-    public float getV2(){
-        return (float) (-b - Math.sqrt(getDiscriminant()))/(2*a);
+    public double getV2(){
+        return (double) (-b - Math.sqrt(getDiscriminant()))/(2*a);
     }
 }
